refactor langchain to expose internal function
